Commit AITK for Python Library at 2025-05-15-14-48-39

+++ Graphics/Color/HSVf.py
@@ -13,7 +13,7 @@ class HSVf                    ( CommonColor                                ) :
   def __init__                ( self                                       ) :
     ##########################################################################
     self . ColorModel = "HSV"
-    self . ValueType  = "Float"
+    self . ValueType  = "Double"
     ##########################################################################
     self . H          = float ( 0.0                                          )
     self . S          = float ( 0.0                                          )
