+++ Graphics/Color/RGBAf.py
@@ -13,7 +13,7 @@ class RGBAf                   ( CommonColor                                ) :
   def __init__                ( self                                       ) :
     ##########################################################################
     self . ColorModel = "RGBA"
-    self . ValueType  = "Float"
+    self . ValueType  = "Double"
     ##########################################################################
     self . R          = float ( 0.0                                          )
     self . G          = float ( 0.0                                          )
